feat: 세부 퀘스트 카운트 타입 Int로 변경

+++ src/main/java/todayquest/quest/dto/DetailInteractRequest.kt
@@ -3,7 +3,7 @@ package todayquest.quest.dto
 import org.hibernate.validator.constraints.Range
 
 class DetailInteractRequest(
-    count: Short? = null
+    count: Int? = null
 ){
     @Range(min = 1, max = 255, message = "{Range.details.count}")
     var count = count ?: 1
